Re-identation and authAPI to validate token
