+++ src/api/ProjectAPI.ts
@@ -8,6 +8,7 @@ import { ProjectController } from "../controller/ProjectController";
 import { UserTypes } from "../entity/UserType";
 import { Project } from "../entity/Project";
 import { Route } from "../config/route";
+import { authApp } from "./AuthAPI";
 
 export const projectApp = express();
 projectApp.use(bodyParser.urlencoded({ extended: true }));
@@ -19,73 +20,46 @@ const userController= new UserController
 const route = new Route
 
 //region requests to PROJECT
-projectApp.post(route.getProjectRoute()+'/allByOwner', async (req, res, next) => {
-  const token = req.headers['x-access-token']
-  if (token) {
-    const validToken = sessionController.validateToken(token.toString())
-    if(validToken!=undefined){
-      asyncConnection().then(async () => {
-        const user = await userController.getUserByEmail(validToken.body.email)
-        res.send(await projectController.getProjectsWithSkills(user))
-      })
-    } else {
-      res.status(401).send("Invalid Token")
-    }
-  } else{
-    res.status(401).send("Pass some Token in header")
-  }
+projectApp.post(route.getProjectRoute()+'/allByOwner', authApp, async (req, res, next) => {
+    const validToken = sessionController.validateToken(req.headers['x-access-token'].toString())
+    asyncConnection().then(async () => {
+    const user = await userController.getUserByEmail(validToken.body.email)
+    res.send(await projectController.getProjectsWithSkills(user))
+    })
 });
   
-projectApp.post(route.getProjectRoute()+'/all', async (req, res, next) => {
-  asyncConnection().then(async () => {
+projectApp.post(route.getProjectRoute()+'/all', authApp, async (req, res, next) => {
+    asyncConnection().then(async () => {
     res.send(await projectController.getProjectsWithSkills(undefined))
-  })
+    })
 });
 
 
-projectApp.post(route.getProjectRoute()+'/insert', async (req, res, next) => {
-  const token = req.headers['x-access-token']
-  if (token) {
-    const validToken = sessionController.validateToken(token.toString())
-    if(validToken!=undefined){
-      asyncConnection().then(async () => {    
-        const user = await userController.getUserByEmail(validToken.body.email.toString())
-        if(user.userType === UserTypes.COMPANY){
-          const project = new Project
-          project.newProject(
-            req.body.name,
-            req.body.description,
-            req.body.closed,
-            user
-          )
-          await projectController.addProject(project)
-          res.send("Project successfully created")
-        } else {
-          res.status(403).send("You cannot create a project since you aren't a company")
-        }
-      })
+projectApp.post(route.getProjectRoute()+'/insert', authApp, async (req, res, next) => {
+    const validToken = sessionController.validateToken(req.headers['x-access-token'].toString())
+    asyncConnection().then(async () => {    
+    const user = await userController.getUserByEmail(validToken.body.email.toString())
+    if(user.userType === UserTypes.COMPANY){
+        const project = new Project
+        project.newProject(
+        req.body.name,
+        req.body.description,
+        req.body.closed,
+        user
+        )
+        await projectController.addProject(project)
+        res.send("Project successfully created")
     } else {
-      res.status(401).send("Invalid Token")
+        res.status(403).send("You cannot create a project since you aren't a company")
     }
-  } else{
-    res.status(401).send("Pass some Token in header")
-  }
+    })
 });
   
-projectApp.post(route.getProjectRoute()+'/skills', async (req, res, next) => {
-  const token = req.headers['x-access-token']
-  if (token) {
-    const validToken = sessionController.validateToken(token.toString())
-    if(validToken!=undefined){
-      asyncConnection().then(async () => {    
+projectApp.post(route.getProjectRoute()+'/skills', authApp, async (req, res, next) => {
+    const validToken = sessionController.validateToken(req.headers['x-access-token'].toString())
+    asyncConnection().then(async () => {    
         const user = await userController.getUserByEmail(validToken.body.email.toString())
         res.send(await projectController.getProjectsWithSkills(user))
-      })
-    } else {
-      res.status(401).send("Invalid Token")
-    }
-  } else{
-    res.status(401).send("Pass some Token in header")
-  }
+    })
 });
